backend/monitoring/requirements.txt
Flask
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
